feat: improve typescript exports
