+++ packages/forge/src/adapters/event.ts
@@ -5,14 +5,11 @@ import type {
   AdapterFactory,
   BaseEventAdapter,
   SanitizedEventConfig,
-  EventOptions,
   EventCallback,
-  EmitOptions,
 } from "@tauriforge/forge/adapters/types";
 
 export class EventAdapter implements BaseEventAdapter {
   readonly name = "event";
-  readonly version = "1.0.0";
 
   private config!: SanitizedEventConfig;
 
@@ -29,16 +26,11 @@ export class EventAdapter implements BaseEventAdapter {
   async on<T = unknown>(
     key: string,
     callback: EventCallback<T>,
-    _options: EventOptions = {},
   ): Promise<UnlistenFn> {
     return await listen<T>(key, (e) => callback(e.payload));
   }
 
-  async emit<T = unknown>(
-    key: string,
-    data: T,
-    _options: EmitOptions = {},
-  ): Promise<void> {
+  async emit<T = unknown>(key: string, data: T): Promise<void> {
     return await emit(key, data);
   }
 }
@@ -48,7 +40,6 @@ export const EventAdapterFactory: AdapterFactory<
   SanitizedEventConfig
 > = {
   name: "event",
-  version: "0.0.1",
   async create(_: SanitizedEventConfig): Promise<EventAdapter> {
     return new EventAdapter();
   },
