+++ packages/forge/src/adapters/index.ts
@@ -1,4 +1,24 @@
-export * from "@tauriforge/forge/adapters/command";
-export * from "@tauriforge/forge/adapters/event";
-export * from "@tauriforge/forge/adapters/registry";
-export * from "@tauriforge/forge/adapters/types";
+export {
+  CommandAdapter,
+  CommandAdapterFactory,
+} from "@tauriforge/forge/adapters/command";
+
+export {
+  EventAdapter,
+  EventAdapterFactory,
+} from "@tauriforge/forge/adapters/event";
+
+export { AdapterRegistry } from "@tauriforge/forge/adapters/registry";
+
+export type {
+  BaseAdapterConfig,
+  BaseAdapter,
+  AdapterFactory,
+  EventAdapterConfig,
+  CommandAdapterConfig,
+  SanitizedCommandConfig,
+  SanitizedEventConfig,
+  EventCallback,
+  BaseEventAdapter,
+  BaseCommandAdapter,
+} from "@tauriforge/forge/adapters/types";
