+++ packages/forge/src/adapters/types.ts
@@ -2,7 +2,6 @@ import type { UnlistenFn } from "@tauri-apps/api/event";
 
 export interface BaseAdapter {
   readonly name: string;
-  readonly version: string;
 
   initialize(config: unknown): Promise<void>;
   destroy(): Promise<void>;
@@ -14,7 +13,6 @@ export interface AdapterFactory<
   TConfig = unknown,
 > {
   name: string;
-  version: string;
   create(config: TConfig): Promise<TAdapter>;
   validateConfig?(config: TConfig): boolean;
 }
@@ -43,11 +41,14 @@ export interface SanitizedEventConfig
 }
 
 export interface BaseCommandAdapter extends BaseAdapter {
-  query<T = unknown>(key: string, options?: QueryOptions): Promise<T>;
+  query<T = unknown>(
+    key: string,
+    // options?: QueryOptions
+  ): Promise<T>;
   mutate<T = unknown, P = unknown>(
     key: string,
     payload?: P,
-    options?: MutationOptions,
+    // options?: MutationOptions
   ): Promise<T>;
 }
 
@@ -55,73 +56,77 @@ export interface BaseEventAdapter extends BaseAdapter {
   on<T = unknown>(
     key: string,
     callback: EventCallback<T>,
-    options?: EventOptions,
+    // options?: EventOptions,
   ): Promise<UnlistenFn>;
 
-  emit<T = unknown>(key: string, data: T, options?: EmitOptions): Promise<void>;
+  emit<T = unknown>(
+    key: string,
+    data: T,
+    // options?: EmitOptions
+  ): Promise<void>;
 }
 
-export interface QueryOptions {
-  timeout?: number;
-  cache?: boolean;
-  retries?: number;
-}
+// export interface QueryOptions {
+//   timeout?: number;
+//   cache?: boolean;
+//   retries?: number;
+// }
 
-export interface MutationOptions extends QueryOptions {
-  optimistic?: boolean;
-  rollback?: boolean;
-}
+// export interface MutationOptions extends QueryOptions {
+//   optimistic?: boolean;
+//   rollback?: boolean;
+// }
 
-export interface EventOptions {
-  once?: boolean;
-  priority?: number;
-}
+// export interface EventOptions {
+//   once?: boolean;
+//   priority?: number;
+// }
 
-export interface EmitOptions {
-  broadcast?: boolean;
-  persist?: boolean;
-}
+// export interface EmitOptions {
+//   broadcast?: boolean;
+//   persist?: boolean;
+// }
 
-export interface BatchQuery {
-  id: string;
-  key: string;
-  options?: QueryOptions;
-}
+// export interface BatchQuery {
+//   id: string;
+//   key: string;
+//   options?: QueryOptions;
+// }
 
-export interface BatchMutation {
-  id: string;
-  key: string;
-  payload?: unknown;
-  options?: MutationOptions;
-}
+// export interface BatchMutation {
+//   id: string;
+//   key: string;
+//   payload?: unknown;
+//   options?: MutationOptions;
+// }
 
-export interface BatchResult<T = unknown> {
-  id: string;
-  success: boolean;
-  data?: T;
-  error?: string;
-}
+// export interface BatchResult<T = unknown> {
+//   id: string;
+//   success: boolean;
+//   data?: T;
+//   error?: string;
+// }
 
 export type EventCallback<T = unknown> = (
   data: T,
   // metadata: EventMetadata
 ) => void;
 
-export interface EventMetadata {
-  timestamp: number;
-  source: string;
-  sequence?: number;
-}
+// export interface EventMetadata {
+//   timestamp: number;
+//   source: string;
+//   sequence?: number;
+// }
 
-export interface BufferedEvent {
-  key: string;
-  data: unknown;
-  timestamp: number;
-  options: EmitOptions;
-}
+// export interface BufferedEvent {
+//   key: string;
+//   data: unknown;
+//   timestamp: number;
+//   options: EmitOptions;
+// }
 
-export interface CacheEntry {
-  data: unknown;
-  createdAt: number;
-  expiresAt: number;
-}
+// export interface CacheEntry {
+//   data: unknown;
+//   createdAt: number;
+//   expiresAt: number;
+// }
