+++ packages/ui/.gitignore
@@ -1,0 +1,2 @@
+# typescript
+*.tsbuildinfo
